Move run classes for analysis
